setting up demo mode
