+++ .env.template
@@ -1,3 +1,4 @@
+DEMO_MODE=true
 CSO_ACCOUNTFACADE_URI=
 CSO_ACCOUNTFACADE_USERNAME=
 CSO_ACCOUNTFACADE_PASSWORD=
