+++ docker-compose.yml
@@ -44,6 +44,7 @@ services:
       - REDIS_PASSWORD=admin
       - NAVIGATION_BASE_URL=http://localhost:3000/efiling
       - NAVIGATION_EXPIRYTIME=10
+      - DEMO_MODE=${DEMO_MODE}
       - CSO_ACCOUNTFACADE_USERNAME=${CSO_ACCOUNTFACADE_USERNAME}
       - CSO_ACCOUNTFACADE_PASSWORD=${CSO_ACCOUNTFACADE_PASSWORD}
       - CSO_ACCOUNTFACADE_URI=${CSO_ACCOUNTFACADE_URI}
