+++ src/backend/efiling-api/src/main/java/ca/bc/gov/open/jag/efilingapi/config/ApplicationProperties.java
@@ -1,0 +1,18 @@
+package ca.bc.gov.open.jag.efilingapi.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "jag.efiling.global")
+public class ApplicationProperties {
+
+    public boolean demo;
+
+    public boolean isDemo() {
+        return demo;
+    }
+
+    public void setDemo(boolean demo) {
+        this.demo = demo;
+    }
+
+}
