--- src/backend/efiling-api/src/main/java/ca/bc/gov/open/jag/efilingapi/config/GlobalCorsConfiguration.java
@@ -1,22 +1,0 @@
-package ca.bc.gov.open.jag.efilingapi.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class GlobalCorsConfiguration {
-/**
-* Configures CORS for enabling javascript applications to connect.
-*/
-    @Bean
-    public WebMvcConfigurer corsConfigurer() {
-        return new WebMvcConfigurer() {
-            @Override
-            public void addCorsMappings(CorsRegistry registry) {
-                registry.addMapping("/**").allowedMethods("*").allowedOrigins("*");
-            }
-        };
-    }
-}
