--- src/backend/efiling-api/src/main/java/ca/bc/gov/open/jag/efilingapi/config/JaegerConfig.java
@@ -1,16 +1,0 @@
-package ca.bc.gov.open.jag.efilingapi.config;
-
-import io.jaegertracing.Configuration;
-import io.opentracing.Tracer;
-import org.springframework.context.annotation.Bean;
-
-
-
-public class JaegerConfig {
-    @Bean
-    public Tracer jaegerTracer() {
-        return new Configuration("spring-boot")
-                .getTracer();
-    }
-
-}
