+++ src/backend/efiling-api/src/test/java/ca/bc/gov/open/jag/efilingapi/config/DemoConfigurationTest.java
@@ -1,21 +1,24 @@
 package ca.bc.gov.open.jag.efilingapi.config;
 
+import ca.bc.gov.open.jag.efilingaccountclient.EfilingAccountService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@DisplayName("JaegerConfig Test Suite")
-public class JaegerConfigTest {
+public class DemoConfigurationTest {
+
+
     ApplicationContextRunner context = new ApplicationContextRunner()
             .withUserConfiguration(ApplicationConfiguration.class);
 
     @Test
-    @DisplayName("CASE1: Test that beans are created")
+    @DisplayName("Test that the demo implementation of the account service is register")
     public void testBeansAreGenerated() {
-        context.run(it -> {
-            assertThat(it).hasSingleBean(ApplicationConfiguration.class);
+        context.withPropertyValues("jag.efiling.global.demo=true").run(it -> {
+            assertThat(it).hasSingleBean(EfilingAccountService.class);
         });
     }
+
 }
